requirements.txt
Flask==2.0.2
tensorflow==2.8.0
numpy==1.22.0
Pillow==9.0.0
requests==2.26.0
opencv-python-headless==4.5.5.62
pytesseract==0.3.8
gunicorn
